Arrabbiata: move challenge semantic into its own module
Simple move, only as the code is larger than initially, and should not be
related to the column module.
